requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
google-adk>=0.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.4.2
websockets>=11.0.3
